+++ src/card.tsx
@@ -1,7 +1,12 @@
 import React from 'react';
-import Badge from './badge';
-import { color as ValidColor } from './types/color';
+
 import cn from './utilities/classnames';
+
+import Badge from './badge';
+import Icon from './icon';
+
+import { color as ValidColor } from './types/color';
+import { IconType } from './types/icons';
 
 export interface CardProps {
   children?: React.ReactNode;
@@ -17,6 +22,9 @@ export interface CardProps {
         color: ValidColor;
       }
     | React.ReactElement<any>;
+  icon?: IconType;
+  iconSize?: IconSizeType;
+  selected?: boolean;
 }
 
 export const Card: React.SFC<CardProps> = ({
@@ -24,10 +32,13 @@ export const Card: React.SFC<CardProps> = ({
   body,
   centered,
   children,
+  icon,
+  iconSize,
   inline,
   thin,
   title,
   className,
+  selected,
   ...attributes,
 }) => {
 
@@ -36,11 +47,17 @@ export const Card: React.SFC<CardProps> = ({
       className={cn('card', className, {
         'is-centered': centered,
         'is-inline': inline,
+        'is-selected': selected,
         'is-thin': thin,
       })}
       {...attributes}
     >
       {badge && <Badge {...badge} />}
+      {icon &&
+        <div className="card-icon">
+          <Icon size={iconSize} type={icon} />
+        </div>
+      }
       {title && <h2 className={cn({ 'card-title': inline })}>{title}</h2>}
       {body && <p>{body}</p>}
       {children}