+++ src/types/icon-sizes.d.ts
@@ -1,0 +1,5 @@
+type IconSizeType =
+  | 20
+  | 30
+  | 40
+  | 60